Made debug hotkeys only appear when the option is enabled

+++ src/engine/game/battle.lua
@@ -3015,14 +3015,16 @@ function Battle:drawDebug()
     self:debugPrintOutline("State: " .. self.state, 4, 0)
     self:debugPrintOutline("Substate: " .. self.substate, 4, 16)
 
-    self:debugPrintOutline("- KEYS -", 4, 64)
-    self:debugPrintOutline("CTRL+H - heal party", 4, 80)
-    self:debugPrintOutline("CTRL+Y - win battle", 4, 96)
-    self:debugPrintOutline("CTRL+M - pause/resume music", 4, 112)
-    self:debugPrintOutline("CTRL+F - end current wave", 4, 128)
-    self:debugPrintOutline("CTRL+B - kill party", 4, 144)
-    self:debugPrintOutline("CTRL+K - fill tension", 4, 160)
-    self:debugPrintOutline("CTRL+N - toggle noclip", 4, 176)
+    if Kristal.Config["debug"] then
+        self:debugPrintOutline("- KEYS -", 4, 64)
+        self:debugPrintOutline("CTRL+H - heal party", 4, 80)
+        self:debugPrintOutline("CTRL+Y - win battle", 4, 96)
+        self:debugPrintOutline("CTRL+M - pause/resume music", 4, 112)
+        self:debugPrintOutline("CTRL+F - end current wave", 4, 128)
+        self:debugPrintOutline("CTRL+B - kill party", 4, 144)
+        self:debugPrintOutline("CTRL+K - fill tension", 4, 160)
+        self:debugPrintOutline("CTRL+N - toggle noclip", 4, 176)
+    end
 end
 
 function Battle:draw()
